NSThread bugfix in sleeping
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@3591 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSThread.m
@@ -196,9 +196,9 @@ void gnustep_base_thread_callback()
   while (delay > 0)
     {
 #ifdef	HAVE_USLEEP
-      usleep (delay*1000000);
+      usleep ((int)(delay*1000000));
 #else
-      sleep (delay);
+      sleep ((int)delay);
 #endif
       delay = [date timeIntervalSinceNow];
     }
